backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
python-dotenv
psycopg2-binary
